gantry changes for SPI refactor
